# docs/openapi-v3.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API para busca textual e vetorial usando Typesense e embeddings gerados via Google Gemini",
    "title": "Mecanismo de Busca API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Prefeitura do Rio de Janeiro",
      "url": "https://prefeitura.rio",
      "email": "contato@prefeitura.rio"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/admin/services/{id}/rollback": {
      "post": {
        "description": "Cria uma nova versão que restaura o estado de uma versão anterior (git-revert style)",
        "tags": [
          "versions"
        ],
        "summary": "Realiza rollback de um serviço para uma versão anterior",
        "parameters": [
          {
            "description": "ID do serviço",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.RollbackRequest"
              }
            }
          },
          "description": "Dados do rollback",
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.PrefRioService"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/admin/services/{id}/versions": {
      "get": {
        "description": "Retorna o histórico completo de versões de um serviço com paginação",
        "tags": [
          "versions"
        ],
        "summary": "Lista todas as versões de um serviço",
        "parameters": [
          {
            "description": "ID do serviço",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Página",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "description": "Resultados por página",
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.VersionHistory"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/services/{id}/versions/compare": {
      "get": {
        "description": "Retorna as diferenças entre duas versões",
        "tags": [
          "versions"
        ],
        "summary": "Compara duas versões de um serviço",
        "parameters": [
          {
            "description": "ID do serviço",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Versão de origem",
            "name": "from_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Versão de destino",
            "name": "to_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.VersionDiff"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/services/{id}/versions/{version}": {
      "get": {
        "description": "Retorna os detalhes de uma versão específica",
        "tags": [
          "versions"
        ],
        "summary": "Busca uma versão específica de um serviço",
        "parameters": [
          {
            "description": "ID do serviço",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Número da versão",
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.ServiceVersion"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://services.pref.rio/app-busca-search",
      "description": "Production server"
    },
    {
      "url": "https://services.staging.app.dados.rio/app-busca-search",
      "description": "Staging server"
    }
  ],
  "components": {
    "schemas": {
      "models.AgentsConfig": {
        "type": "object",
        "properties": {
          "exclusive_for_agents": {
            "type": "boolean"
          },
          "tool_hint": {
            "type": "string"
          }
        }
      },
      "models.Button": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string"
          },
          "is_enabled": {
            "type": "boolean"
          },
          "ordem": {
            "type": "integer"
          },
          "titulo": {
            "type": "string"
          },
          "url_service": {
            "type": "string"
          }
        }
      },
      "models.FieldChange": {
        "type": "object",
        "properties": {
          "field_name": {
            "type": "string",
            "maxLength": 20000
          },
          "new_value": {},
          "old_value": {},
          "value_type": {
            "description": "\"string\", \"int\", \"bool\", \"array\", \"object\"",
            "type": "string",
            "maxLength": 20000
          }
        }
      },
      "models.PrefRioService": {
        "type": "object",
        "required": [
          "autor",
          "custo_servico",
          "descricao_completa",
          "nome_servico",
          "orgao_gestor",
          "resultado_solicitacao",
          "resumo",
          "tema_geral",
          "tempo_atendimento"
        ],
        "properties": {
          "agents": {
            "$ref": "#/components/schemas/models.AgentsConfig"
          },
          "autor": {
            "type": "string",
            "maxLength": 20000
          },
          "awaiting_approval": {
            "type": "boolean"
          },
          "buttons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.Button"
            }
          },
          "canais_digitais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canais_presenciais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "integer"
          },
          "custo_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "descricao_completa": {
            "type": "string",
            "maxLength": 20000
          },
          "documentos_necessarios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "embedding": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "extra_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "fixar_destaque": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "instrucoes_solicitante": {
            "type": "string",
            "maxLength": 20000
          },
          "is_free": {
            "type": "boolean"
          },
          "last_update": {
            "type": "integer"
          },
          "legislacao_relacionada": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nome_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "orgao_gestor": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string"
            }
          },
          "publico_especifico": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published_at": {
            "type": "integer"
          },
          "resultado_solicitacao": {
            "type": "string",
            "maxLength": 20000
          },
          "resumo": {
            "type": "string",
            "maxLength": 20000
          },
          "search_content": {
            "type": "string"
          },
          "servico_nao_cobre": {
            "type": "string",
            "maxLength": 20000
          },
          "status": {
            "description": "0=Draft, 1=Published",
            "type": "integer",
            "maximum": 1,
            "minimum": 0
          },
          "tema_geral": {
            "type": "string",
            "maxLength": 20000
          },
          "tempo_atendimento": {
            "type": "string",
            "maxLength": 20000
          }
        }
      },
...
      "models.RollbackRequest": {
        "type": "object",
        "required": [
          "to_version"
        ],
        "properties": {
          "change_reason": {
            "type": "string",
            "maxLength": 20000
          },
          "to_version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "models.ServiceVersion": {
        "type": "object",
        "required": [
          "change_type",
          "created_by",
          "created_by_cpf",
          "service_id",
          "version_number"
        ],
        "properties": {
          "autor": {
            "type": "string",
            "maxLength": 20000
          },
          "awaiting_approval": {
            "type": "boolean"
          },
          "canais_digitais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canais_presenciais": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "change_reason": {
            "type": "string",
            "maxLength": 20000
          },
          "change_type": {
            "type": "string",
            "enum": [
              "create",
              "update",
              "publish",
              "unpublish",
              "delete",
              "rollback"
            ]
          },
          "changed_fields_json": {
            "description": "Campos de mudança (armazenados como JSON string no Typesense)",
            "type": "string",
            "maxLength": 20000
          },
          "created_at": {
            "type": "integer"
          },
          "created_by": {
            "type": "string",
            "maxLength": 20000
          },
          "created_by_cpf": {
            "type": "string",
            "maxLength": 20000
          },
          "custo_servico": {
            "type": "string",
            "maxLength": 20000
          },
          "descricao_completa": {
            "type": "string",
            "maxLength": 20000
          },
          "documentos_necessarios": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "embedding_hash": {
            "description": "Hash do embedding para verificação (não armazenamos o embedding completo)",
            "type": "string",
            "maxLength": 20000
          },
          "fixar_destaque": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "instrucoes_solicitante": {
            "type": "string",
            "maxLength": 20000
          },
          "is_free": {
            "type": "boolean"
          },
          "is_rollback": {
            "type": "boolean"
          },
          "legislacao_relacionada": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nome_servico": {
            "description": "Snapshot completo do serviço (sem embedding para economizar espaço)",
            "type": "string",
            "maxLength": 20000
          },
          "orgao_gestor": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "previous_version": {
            "type": "integer"
          },
          "publico_especifico": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published_at": {
            "type": "integer"
          },
          "resultado_solicitacao": {
            "type": "string",
            "maxLength": 20000
          },
          "resumo": {
            "type": "string",
            "maxLength": 20000
          },
          "rollback_to_version": {
            "type": "integer"
          },
          "search_content": {
            "type": "string",
            "maxLength": 20000
          },
          "service_id": {
            "type": "string",
            "maxLength": 20000
          },
          "servico_nao_cobre": {
            "type": "string",
            "maxLength": 20000
          },
          "status": {
            "type": "integer"
          },
          "tema_geral": {
            "type": "string",
            "maxLength": 20000
          },
          "tempo_atendimento": {
            "type": "string",
            "maxLength": 20000
          },
          "version_number": {
            "type": "integer"
          }
        }
      },
...
      "models.VersionDiff": {
        "type": "object",
        "properties": {
          "change_type": {
            "type": "string"
          },
          "changed_at": {
            "type": "integer"
          },
          "changed_by": {
            "type": "string"
          },
          "changes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.FieldChange"
            }
          },
          "from_version": {
            "type": "integer"
          },
          "to_version": {
            "type": "integer"
          }
        }
      },
      "models.VersionHistory": {
        "type": "object",
        "properties": {
          "found": {
            "type": "integer"
          },
          "out_of": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/models.ServiceVersion"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}